Add additional unit tests
