+++ tests/test_add_classifications_stage_pipe.py
@@ -19,6 +19,10 @@ import os
 import numpy as np
 import pandas as pd
 
+from morpheus.messages import MessageMeta
+from morpheus.messages import MultiInferenceMessage
+from morpheus.messages import MultiMessage
+from morpheus.messages import MultiResponseProbsMessage
 from morpheus.pipeline import LinearPipeline
 from morpheus.stages.input.file_source_stage import FileSourceStage
 from morpheus.stages.output.write_to_file_stage import WriteToFileStage
@@ -45,6 +49,45 @@ def test_add_classifications_stage_pipe(config, tmp_path):
     pipe.add_stage(ConvMsg(config, input_file))
     pipe.add_stage(AddClassificationsStage(config, threshold=threshold))
     pipe.add_stage(SerializeStage(config, include=["^{}$".format(c) for c in config.class_labels]))
+    pipe.add_stage(WriteToFileStage(config, filename=out_file, overwrite=False))
+    pipe.run()
+
+    assert os.path.exists(out_file)
+
+    input_data = np.loadtxt(input_file, delimiter=",", skiprows=1)
+    expected = (input_data > threshold)
+
+    # The output data will contain an additional id column that we will need to slice off
+    output_data = pd.read_csv(out_file)
+    idx = output_data.columns.intersection(config.class_labels)
+    assert idx.to_list() == config.class_labels
+
+    output_np = output_data[idx].to_numpy()
+
+    assert output_np.tolist() == expected.tolist()
+
+
+def test_add_classifications_stage_multi_segment_pipe(config, tmp_path):
+    config.class_labels = ['frogs', 'lizards', 'toads', 'turtles']
+    config.num_threads = 1
+
+    # Silly data with all false values
+    input_file = os.path.join(TEST_DIRS.tests_data_dir, "filter_probs.csv")
+    out_file = os.path.join(tmp_path, 'results.csv')
+
+    threshold = 0.75
+
+    pipe = LinearPipeline(config)
+    pipe.set_source(FileSourceStage(config, filename=input_file, iterative=False))
+    pipe.add_segment_boundary(MessageMeta)
+    pipe.add_stage(DeserializeStage(config))
+    pipe.add_segment_boundary(MultiMessage)
+    pipe.add_stage(ConvMsg(config, input_file))
+    pipe.add_segment_boundary(MultiResponseProbsMessage)
+    pipe.add_stage(AddClassificationsStage(config, threshold=threshold))
+    pipe.add_segment_boundary(MultiResponseProbsMessage)
+    pipe.add_stage(SerializeStage(config, include=["^{}$".format(c) for c in config.class_labels]))
+    pipe.add_segment_boundary(MessageMeta)
     pipe.add_stage(WriteToFileStage(config, filename=out_file, overwrite=False))
     pipe.run()
