+++ tests/test_filter_detections_stage_pipe.py
@@ -18,14 +18,18 @@ import os
 
 import numpy as np
 
+from morpheus.messages import MessageMeta
+from morpheus.messages import MultiInferenceMessage
+from morpheus.messages import MultiMessage
+from morpheus.messages import MultiResponseProbsMessage
 from morpheus.pipeline import LinearPipeline
 from morpheus.stages.input.file_source_stage import FileSourceStage
 from morpheus.stages.output.write_to_file_stage import WriteToFileStage
 from morpheus.stages.postprocess.filter_detections_stage import FilterDetectionsStage
 from morpheus.stages.postprocess.serialize_stage import SerializeStage
 from morpheus.stages.preprocess.deserialize_stage import DeserializeStage
-from utils import TEST_DIRS
 from utils import ConvMsg
+from utils import TEST_DIRS
 
 
 def _test_filter_detections_stage_pipe(config, tmp_path, copy=True):
@@ -56,9 +60,50 @@ def _test_filter_detections_stage_pipe(config, tmp_path, copy=True):
     assert output_data.tolist() == expected.tolist()
 
 
+def test_filter_detections_stage_multi_segment_pipe(config, tmp_path, copy=True):
+    input_file = os.path.join(TEST_DIRS.tests_data_dir, "filter_probs.csv")
+    out_file = os.path.join(tmp_path, 'results.csv')
+
+    threshold = 0.75
+
+    pipe = LinearPipeline(config)
+    pipe.set_source(FileSourceStage(config, filename=input_file, iterative=False))
+    pipe.add_segment_boundary(MessageMeta)
+    pipe.add_stage(DeserializeStage(config))
+    pipe.add_segment_boundary(MultiMessage)
+    pipe.add_stage(ConvMsg(config))
+    pipe.add_segment_boundary(MultiResponseProbsMessage)
+    pipe.add_stage(FilterDetectionsStage(config, threshold=threshold, copy=copy))
+    pipe.add_segment_boundary(MultiResponseProbsMessage)
+    pipe.add_stage(SerializeStage(config))
+    pipe.add_segment_boundary(MessageMeta)
+    pipe.add_stage(WriteToFileStage(config, filename=out_file, overwrite=False))
+    pipe.run()
+
+    assert os.path.exists(out_file)
+
+    input_data = np.loadtxt(input_file, delimiter=",", skiprows=1)
+    output_data = np.loadtxt(out_file, delimiter=",", skiprows=1)
+
+    # The output data will contain an additional id column that we will need to slice off
+    # also somehow 0.7 ends up being 0.7000000000000001
+    output_data = np.around(output_data[:, 1:], 2)
+
+    expected = input_data[np.any(input_data >= threshold, axis=1), :]
+    assert output_data.tolist() == expected.tolist()
+
+
 def test_filter_detections_stage_pipe_copy(config, tmp_path):
     return _test_filter_detections_stage_pipe(config, tmp_path, True)
 
 
 def test_filter_detections_stage_pipe_slice(config, tmp_path):
     return _test_filter_detections_stage_pipe(config, tmp_path, False)
+
+def test_filter_detections_stage_multi_segment_pipe_copy(config, tmp_path):
+    return _test_filter_detections_stage_pipe(config, tmp_path, True)
+
+
+def test_filter_detections_stage_multi_segment_pipe_slice(config, tmp_path):
+    return _test_filter_detections_stage_pipe(config, tmp_path, False)
+
